Changed rates to be integers instead of doubles for random function to be easier to use, added random function to attack, added applyCrit() to Attack & added messages for the case an attack would crit,miss, or be evaded

+++ src/main/java/Character.java
@@ -8,7 +8,7 @@ public class Character extends Battler{
     }
 
     public Character(String name, int HP, int MaxHP, int MP, int MaxMP, int Atk, int Def,
-                     double CritRate, double HitRate, double EvaRate){
+                     int CritRate, int HitRate, int EvaRate){
         super(name, HP, MaxHP, MP, MaxMP, Atk, Def, CritRate, HitRate, EvaRate);
         addSpecialAttack(new DualSlash());
         addSpecialAttack(new Pierce());
