+++ src/main/java/Enemy.java
@@ -6,7 +6,7 @@ public class Enemy extends Battler{
     }
 
     public Enemy(String name, int HP, int MaxHP, int MP, int MaxMP, int Atk, int Def,
-                     double CritRate, double HitRate, double EvaRate) {
+                     int CritRate, int HitRate, int EvaRate) {
         super(name, HP, MaxHP, MP, MaxMP, Atk, Def, CritRate, HitRate, EvaRate);
         addSpecialAttack(new Bash());
     }
